Refactor com_atproto::sync
