+++ atprotolib-rs/src/types/com_atproto/mod.rs
@@ -16,30 +16,4 @@ pub mod repo;
 pub mod server;
 
 /// Contains types for the `com.atproto.sync` namespace.
-pub mod sync {
-    mod get_head;
-    mod get_latest_commit;
-    mod get_repo_status;
-    mod list_blobs;
-    mod list_repos;
-    mod notify_of_update;
-    mod request_crawl;
-    mod subscribe_repos;
-
-    pub use self::{
-        get_head::*,
-        get_latest_commit::*,
-        get_repo_status::*,
-        list_blobs::*,
-        list_repos::*,
-        notify_of_update::*,
-        request_crawl::*,
-        subscribe_repos::*
-    };
-
-    #[cfg(feature = "apicalls")]
-    mod api_calls;
-
-    #[cfg(feature = "apicalls")]
-    pub use self::api_calls::*;
-}
+pub mod sync;
