+++ atprotolib-rs/src/types/com_atproto/sync/api_calls.rs
@@ -1,21 +1,15 @@
-use crate::{
-    api_calls::{AddApiAuth, ApiAuthConfig, ApiError},
-    types::com_atproto::sync::{
-        GetLatestCommitResponse,
-        GetRepoStatusResponse,
-        ListBlobsResponse,
-        ListReposResponse,
-        NotifyOfUpdateRequest,
-        RequestCrawlRequest
-    }
-};
+use crate::api_calls::{AddApiAuth, ApiAuthConfig, ApiError};
 
-/// Get a blob associated with a given account. Returns the full blob as originally uploaded. Does not require auth; implemented by PDS.
-/// 
+use super::{api_requests::{CrawlRequest, NotifyOfUpdateRequest}, api_responses::{LatestCommitResponse, ListBlobsResponse, ListReposResponse, RepoStatusResponse}};
+
+/// Get a blob associated with a given account. Returns the full blob as
+/// originally uploaded. Does not require auth; implemented by PDS.
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the server to send the request to.
-/// * `api_auth_config` - The authentication configuration to use for the request.
+/// * `api_auth_config` - The authentication configuration to use for the
+///   request.
 /// * `did` - The DID of the account.
 /// * `cid` - The CID of the blob to get.
 pub async fn get_blob(
@@ -29,7 +23,6 @@ pub async fn get_blob(
 
     let query_params = vec![("did", did), ("cid", cid)];
 
-    
     let response = client
         .get(&api_url)
         .query(&query_params)
@@ -46,19 +39,21 @@ pub async fn get_blob(
     }
 }
 
-/// Get the current commit CID & revision of the specified repo. Does not require auth.
-/// 
+/// Get the current commit CID & revision of the specified repo. Does not
+/// require auth.
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the server to send the request to.
-/// * `api_auth_config` - The authentication configuration to use for the request.
+/// * `api_auth_config` - The authentication configuration to use for the
+///   request.
 /// * `did` - The DID of the repo.
 pub async fn get_latest_commit(
     host_name: &str,
     client: reqwest::Client,
     api_auth_config: &ApiAuthConfig,
     did: &str
-) -> Result<GetLatestCommitResponse, Box<dyn std::error::Error>> {
+) -> Result<LatestCommitResponse, Box<dyn std::error::Error>> {
     let api_url = format!(
         "https://{}/xrpc/com.atproto.sync.getLatestCommit",
         host_name
@@ -66,7 +61,6 @@ pub async fn get_latest_commit(
 
     let query_params = vec![("did", did)];
 
-    
     let response = client
         .get(&api_url)
         .query(&query_params)
@@ -76,19 +70,21 @@ pub async fn get_latest_commit(
 
     match response.status() {
         reqwest::StatusCode::OK => {
-            let response_body: GetLatestCommitResponse = response.json().await?;
+            let response_body: LatestCommitResponse = response.json().await?;
             Ok(response_body)
         }
         _ => Err(Box::new(ApiError::new(response).await?))
     }
 }
 
-/// Get data blocks needed to prove the existence or non-existence of record in the current version of repo. Does not require auth.
-/// 
+/// Get data blocks needed to prove the existence or non-existence of record in
+/// the current version of repo. Does not require auth.
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the server to send the request to.
-/// * `api_auth_config` - The authentication configuration to use for the request.
+/// * `api_auth_config` - The authentication configuration to use for the
+///   request.
 /// * `did` - The DID of the repo.
 /// * `collection` - The collection of the record.
 /// * `rkey` - The rkey of the record.
@@ -104,7 +100,6 @@ pub async fn get_record(
 
     let query_params = vec![("did", did), ("collection", collection), ("rkey", rkey)];
 
-    
     let response = client
         .get(&api_url)
         .query(&query_params)
@@ -121,24 +116,25 @@ pub async fn get_record(
     }
 }
 
-/// Get the hosting status for a repository, on this server. Expected to be implemented by PDS and Relay.
-/// 
+/// Get the hosting status for a repository, on this server. Expected to be
+/// implemented by PDS and Relay.
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the server to send the request to.
-/// * `api_auth_config` - The authentication configuration to use for the request.
+/// * `api_auth_config` - The authentication configuration to use for the
+///   request.
 /// * `did` - The DID of the repo.
 pub async fn get_repo_status(
     host_name: &str,
     client: reqwest::Client,
     api_auth_config: &ApiAuthConfig,
     did: &str
-) -> Result<GetRepoStatusResponse, Box<dyn std::error::Error>> {
+) -> Result<RepoStatusResponse, Box<dyn std::error::Error>> {
     let api_url = format!("https://{}/xrpc/com.atproto.sync.getRepoStatus", host_name);
 
     let query_params = vec![("did", did)];
 
-    
     let response = client
         .get(&api_url)
         .query(&query_params)
@@ -148,19 +144,21 @@ pub async fn get_repo_status(
 
     match response.status() {
         reqwest::StatusCode::OK => {
-            let response_body: GetRepoStatusResponse = response.json().await?;
+            let response_body: RepoStatusResponse = response.json().await?;
             Ok(response_body)
         }
         _ => Err(Box::new(ApiError::new(response).await?))
     }
 }
 
-/// Download a repository export as CAR file. Optionally only a 'diff' since a previous revision. Does not require auth; implemented by PDS.
-/// 
+/// Download a repository export as CAR file. Optionally only a 'diff' since a
+/// previous revision. Does not require auth; implemented by PDS.
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the server to send the request to.
-/// * `api_auth_config` - The authentication configuration to use for the request.
+/// * `api_auth_config` - The authentication configuration to use for the
+///   request.
 /// * `did` - The DID of the repo.
 /// * `since` - The CID of the previous revision to diff against.
 pub async fn get_repo(
@@ -179,7 +177,6 @@ pub async fn get_repo(
         query_params.push(("since", since));
     }
 
-    
     let response = client
         .get(&api_url)
         .query(&query_params)
@@ -196,12 +193,14 @@ pub async fn get_repo(
     }
 }
 
-/// List blob CIDs for an account, since some repo revision. Does not require auth; implemented by PDS.
-/// 
+/// List blob CIDs for an account, since some repo revision. Does not require
+/// auth; implemented by PDS.
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the server to send the request to.
-/// * `api_auth_config` - The authentication configuration to use for the request.
+/// * `api_auth_config` - The authentication configuration to use for the
+///   request.
 /// * `did` - The DID of the account.
 /// * `since` - The CID of the previous revision to diff against.
 /// * `limit` - The maximum number of blobs to return.
@@ -229,7 +228,6 @@ pub async fn list_blobs(
         query_params.push(("cursor", cursor));
     }
 
-    
     let response = client
         .get(&api_url)
         .query(&query_params)
@@ -246,12 +244,14 @@ pub async fn list_blobs(
     }
 }
 
-/// Enumerates all the DID, rev, and commit CID for all repos hosted by this service. Does not require auth; implemented by PDS and Relay.
-/// 
+/// Enumerates all the DID, rev, and commit CID for all repos hosted by this
+/// service. Does not require auth; implemented by PDS and Relay.
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the server to send the request to.
-/// * `api_auth_config` - The authentication configuration to use for the request.
+/// * `api_auth_config` - The authentication configuration to use for the
+///   request.
 /// * `limit` - The maximum number of repos to return.
 /// * `cursor` - The cursor to use for pagination.
 pub async fn list_repos(
@@ -270,7 +270,6 @@ pub async fn list_repos(
         query_params.push(("cursor", cursor));
     }
 
-    
     let response = client
         .get(&api_url)
         .query(&query_params)
@@ -287,12 +286,15 @@ pub async fn list_repos(
     }
 }
 
-/// Notify a crawling service of a recent update, and that crawling should resume. Intended use is after a gap between repo stream events caused the crawling service to disconnect. Does not require auth; implemented by Relay.
-/// 
+/// Notify a crawling service of a recent update, and that crawling should
+/// resume. Intended use is after a gap between repo stream events caused the
+/// crawling service to disconnect. Does not require auth; implemented by Relay.
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the server to send the request to.
-/// * `api_auth_config` - The authentication configuration to use for the request.
+/// * `api_auth_config` - The authentication configuration to use for the
+///   request.
 /// * `request` - The request to send.
 pub async fn notify_of_update(
     host_name: &str,
@@ -302,7 +304,6 @@ pub async fn notify_of_update(
 ) -> Result<(), Box<dyn std::error::Error>> {
     let api_url = format!("https://{}/xrpc/com.atproto.sync.notifyOfUpdate", host_name);
 
-    
     let response = client
         .post(&api_url)
         .json(&request)
@@ -316,22 +317,23 @@ pub async fn notify_of_update(
     }
 }
 
-/// Request a service to persistently crawl hosted repos. Expected use is new PDS instances declaring their existence to Relays. Does not require auth.
-/// 
+/// Request a service to persistently crawl hosted repos. Expected use is new
+/// PDS instances declaring their existence to Relays. Does not require auth.
+///
 /// ## Arguments
-/// 
+///
 /// * `host_name` - The host name of the server to send the request to.
-/// * `api_auth_config` - The authentication configuration to use for the request.
+/// * `api_auth_config` - The authentication configuration to use for the
+///   request.
 /// * `request` - The request to send.
 pub async fn request_crawl(
     host_name: &str,
     client: reqwest::Client,
     api_auth_config: &ApiAuthConfig,
-    request: RequestCrawlRequest
+    request: CrawlRequest
 ) -> Result<(), Box<dyn std::error::Error>> {
     let api_url = format!("https://{}/xrpc/com.atproto.sync.requestCrawl", host_name);
 
-    
     let response = client
         .post(&api_url)
         .json(&request)
