--- atprotolib-rs/src/types/com_atproto/sync/get_head.rs
@@ -1,12 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-/*
-    com.atproto.sync.getHead
-*/
-
-#[allow(missing_docs)]
-#[derive(Serialize, Deserialize, Debug)]
-pub struct GetHeadResponse {
-    #[serde(rename = "root")]
-    pub root: String
-}
