--- atprotolib-rs/src/types/com_atproto/sync/request_crawl.rs
@@ -1,13 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-/*
-    com.atproto.sync.requestCrawl
-*/
-
-/// Represents a request to crawl a hostname.
-#[derive(Serialize, Deserialize, Debug)]
-pub struct RequestCrawlRequest {
-    /// The hostname to crawl.
-    #[serde(rename = "hostname")]
-    pub hostname: String
-}
